Make board file generic and use a trigger file
This is a work in progress. The trigger symbols are marked as weak so the user
can provide a file that implements and overloads them.

Also, we can use the avr-avrdude.exp board file as a generic configuration to
work with avrdude, setting its options in environment variables thanks to
run.avr.sh .

	* dejagnu/baseboards/atmega328p-shrimp.exp: Make all the flags for
	avrdude in a single variable.
	* dejagnu/baseboards/atxmega256a3bu.exp: Likewise. Also add a trigger
	file.
	* dejagnu/baseboards/avr-avrdude.exp: New generic board file. This board
	file takes its values from environment variables set by the user and the
	run-avr.sh script.
	* dejagnu/config/avrdude.exp: Make all the flags for avrdude in a single
	variable.
	* gcc.beebs/platformcode.h: Give the weak attribute to the symbols.
	* run-avr.sh: Allow the user to set environment variables.